chore: move splashscreen into CordovaActivity
