+++ templates/project/Activity.java
@@ -20,26 +20,15 @@
 package __ID__;
 
 import android.os.Bundle;
-import android.os.Handler;
-import android.util.Log;
 
 import org.apache.cordova.*;
-import androidx.core.splashscreen.SplashScreen;
 
 public class __ACTIVITY__ extends CordovaActivity
 {
-    private boolean splashScreenKeepOnScreen = true;
-
     @Override
     public void onCreate(Bundle savedInstanceState)
     {
-        // Handle the splash screen transition.
-        SplashScreen splashScreen = SplashScreen.installSplashScreen(this);
-
         super.onCreate(savedInstanceState);
-
-        // Setup the splash screen delay based on preference settings
-        setupSplashScreenDelay(splashScreen);
 
         // enable Cordova apps to be started in the background
         Bundle extras = getIntent().getExtras();
@@ -49,23 +38,5 @@ public class __ACTIVITY__ extends CordovaActivity
 
         // Set by <content src="index.html" /> in config.xml
         loadUrl(launchUrl);
-    }
-
-    private void setupSplashScreenDelay(SplashScreen splashScreen) {
-        boolean splashScreenAutoHide = preferences.getBoolean("AutoHideSplashScreen", true);
-        int splashScreenDelay = preferences.getInteger("SplashScreenDelay", 3000);
-
-        Log.d("CdvSplashScreen", "Auto Hide: " + splashScreenAutoHide);
-
-        if (splashScreenAutoHide) {
-            Log.d("CdvSplashScreen", "Delay: " + splashScreenDelay + "ms");
-
-            splashScreen.setKeepOnScreenCondition(() -> splashScreenKeepOnScreen);
-            Handler splashScreenDelayHandler = new Handler();
-            splashScreenDelayHandler.postDelayed(
-                () -> splashScreenKeepOnScreen = false,
-                splashScreenDelay
-            );
-        }
     }
 }
