+++ templates/project/app/build.gradle
@@ -24,12 +24,6 @@ if (cordovaConfig.IS_GRADLE_PLUGIN_KOTLIN_ENABLED) {
     apply plugin: 'kotlin-android-extensions'
 }
 
-dependencies {
-    apply from: '../CordovaLib/cordova.gradle'
-
-    implementation "androidx.core:core-splashscreen:${cordovaConfig.ANDROIDX_CORE_SPLASHSCREEN_VERSION}"
-}
-
 buildscript {
     apply from: '../CordovaLib/cordova.gradle'
